requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-appconf==1.0.6
django-compressor==4.4
django-redis==5.4.0
django-sass-processor==1.4
gunicorn==21.2.0
libsass==0.23.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
rcssmin==1.1.1
redis==5.0.1
rjsmin==1.2.1
sqlparse==0.4.4
typing_extensions==4.8.0
